fix: 포인트 제외 디버깅. 프로필 정리

+++ src/main/java/store/buzzbook/core/repository/user/GradeRepository.java
@@ -1,12 +1,12 @@
 package store.buzzbook.core.repository.user;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import store.buzzbook.core.entity.user.Grade;
 import store.buzzbook.core.entity.user.GradeName;
 
-import java.util.Optional;
-
-public interface GradeRepository extends JpaRepository<Grade, Long> {
-    Optional<Grade> findByName(GradeName name);
+public interface GradeRepository extends JpaRepository<Grade, Integer> {
+	Optional<Grade> findByName(GradeName name);
 }
